Simplify tab-state and tab-manager modules
- no more scheduled logging so these can be slimmed down a lot
- migrated them to TS as well
